se rompio todo

+++ Dominio/Especialidad.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
    public class Especialidad
     {
-        public int idEspecialidad { get; set; }
+        public long idEspecialidad { get; set; }
         public string nombre { get; set; }
         public int convenio { get; set; }
         public Byte estado { get; set; }
